Move values to a separate module in the kernel
